Add TCP Session database cleanup
The original TCP session database code was missing a cleanup function to be invoked
when the RevSw modules is removed.  This orphaned any session entry memory that was
still in the databse and caused core dumps.  The cleanup function needs to cancel
all outstanding working items, remove the list from the hash and then free the
memory.  Once this is complete it must delete the hash table itself.

Signed-off-by: Tom Kavanagh <tom.revsw@gmail.com>