Receive Window Multiplier
We are no longer setting the client-side receive window arbitrarily
to 100K and the initial congestion window to 100.  Instead we are
providing the following:

1. Add a receive window multiplier.  This will be used to increase
the initial client-side receive window by a specified factor.

2. There will be a default multiplier for initial receive windows
of less than 64K and a separate mutiplier for windows larger than
64K.

3.  The initial congestion window sysctl will be defaulted to 0 so
that the system will determine the  appropriate initial congestion
window based off of the new receive window and packet size of 1K.
If the sysctl is set to anything other than 0 then this will be used
as the initial congestion window.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>