Renaming CCA ops get_leak_quota() name
Changing get_leak_quota() API name to get_cwnd_quota().  This is done to
keep the code in line with what is seen in the tcp_write_xmit function and
to hide the fact we have implemented a leaky bucket algorithm from any
kernel code we have to make public.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>