Introduce snd_wnd_test() API pointer to CCA ops vector
Added a new API pointer to the congestion control algoritm to allow each
algorithm to provide its own send window test.  This is currently needed
by RRE but not by the RevSw algorithm.  This is also done to make the
avoidance of executing a send window test explicit and not a side affect
of defining a unrelated API pointer.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>