Add a tcp_rre_snd_wnd_test() API
With the introduction of a snd_wnd_test() API pointer to the CCA ops
vector, the RRE algorithm needs to define its version of this API in
order to avoid having the standard TCP API from being invoked.

Also renamed the tcp_revsw_handle_nagle_test() function in tcp_rre.c
to tcp_rre_handle_nagle_test() to keep it consistent with the naming
convention in the file.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>