Adjust RevSw CCA to account for active connections
The initial congestion window was a set value regardless of how many connections
were active to a give IP address.  The could flood that device with data.  So
modifying the algorithm to look at whether or not there are active connections
to the IP address and if so, cut the initial congestion window in half.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>