Removed '/' operator from tcp_revsw_ssthresh()
The usage of arithmatic operators in some of the TCP CCA code will cause
scheduling BUG traces.  While these are not fully understood yet, a fix is
to replace these operators with calls to internal functions that perform
the operation without the use of the operators.  In this case the '/' in
tcp_revsw_ssthresh() is replaced with a call to tcp_revsw_division().

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>