Remove del_timer() call from tcp_session_delete_work_handler()
The tcp_session_delete_work_handler() function is a generic function that should
not have any CCA specific code in it. The call to del_timer() is specific to RBE
and could cause core dumps or memory corruption when the standard RevSw CCA is
used.  So removing it for now and we will need to find a new way of providing CCA
specific handlign from within this function.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>