DOTS-28 Fix module verification fail
The module verification fail occurs becaue the modules are not signed.  This
is usually done when the modules_install make command is executed.  The
RevSw modules are not installed in this manner so the modules have to be
signed manually.  Modified the build kernel script to manually sign the
modules when they are build.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>