DOTS SafetyNet
Modified the TCP Session Database to hold a connections list and a client list.
The client list holds the latest information about a given IP address and will
serve as the session databse.  The connections list serves as a storage base
for all active connections that will be used to update the client list.

Implemented a safety net for DOTS.  The client information that is stored will
be used to determine whether or not DOTS is causing issues for that particular
client.  If so there are various levels of backing off that DOTS will use to
provide the best service possible to the client.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>