Fix TCP RevSw Session database removal
The unregister of the tcp_revsw_session_db was hanging the system.
This was due to the manner in which all of the memory was freed.
Originally all of the entries were individual allocated and freed.
Now they are allocated as a block but the free code was still individually.
Fix it so the free is on a block by block basis.

Signed-off-by: Tom Kavanagh <tom.revsw@gmail.com>