Limiting the maximum number of TCP Session DB client records
Due to memory constraints on the kernel, it is required that the maximum number
of client records be restricted.  Setting the default maximum to 3000 with a
sysctl to change this (max value of 10,000).  Also, the client records are held
only for client that have experienced a certain level of transmission retransmission.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>