Fix NULL pointer dereference in session deletion
It is possible that a socket is being closed that has not had a
session record attached to it (it never reached the connected state).
in this scenario the current session delete API will result in a
kernel crash due to a NULL pointer dereference.  Checking the
session pointer before using it and if NULL then simply return.

Signed-off-by: Thomas Kavanagh <tom.revsw@gmail.com>