DOTS-53 Module name change causes issues with setting net.ipv4.tcp_congestion_control
In order for the RevSw CCA module to be automatically installed when setting the sysctl
and upon reboot, the module name MUST be tcp_revsw.  So making the necessary name changes
to the files and module.

Also adding in a dummy CCA to help track down issues with CCA APIs being called for
sockets that have not had the CCA init function called for them or have had the release
function called.  This will help prevent NULL pointer exceptions.

Signed-off-by: Tom Kavanagh <tom.revsw@gmail.com>